cleaned IPCHandlers
